Replaced WebView with WebView2
Webview has been replaced with a more modern Webview 2 using Chromium engine, instead of (to my knowledge) the legacy EdgeHTML engine.

The minimum target had to be increased from 10.0.16299.0 to 10.0.17763.0 for Webview2. And the Microsoft.UI.Xaml nuget package had to be added as dependency.

This allows the UWP client the render the 10.9.9 Jellyfin web.